# frontend/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pi Control API",
    "description": "面向应用、脚本、家庭自动化和第三方集成的 HTTP / JSON API。\n\n基础路径为 `/api/v1`，时间戳使用 Unix 秒。接口按资源分组，提供参数、响应结构与示例。\n\n**认证**：在控制台「设置 → 开发者 API」创建令牌，通过 `Authorization: Bearer <token>` 调用。\n`read` 可查询及生成 OLED 预览，`control` 可修改硬件配置和场景，`admin` 为网页登录会话，可管理令牌与设备目录。\n当前已登录的浏览器可直接调试；也可点击 Authorize 输入完整令牌（不用再输入 Bearer 前缀）。\n\n**控制请求**：先 GET `/status`，再提交 `expected_version`、唯一 `request_id` 和 `value`。\n`value` 是对应资源的替换配置，省略字段使用默认值；保留其余配置时应基于 GET 返回值修改。\n409 时重新读取并处理冲突；503 时目标配置可能已经持久化，应同时检查 `config`、`applied` 与 `hardware.errors`。\n\n**实时数据**：WebSocket `/api/v1/live` 每约 0.5 秒推送完整状态，详见上方接入指南。\n",
    "version": "1.5.0"
  },
  "paths": {
...
    "/api/v1/device": {
      "get": {
        "summary": "获取公开设备标识",
        "operationId": "device_identity",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceIdentity"
                }
              }
            }
          },
          "400": {
            "description": "Host 不被允许。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "Host 不被允许。"
                }
              }
            }
          }
        },
        "tags": [
          "Status"
        ],
        "description": "**最低权限：无需认证。**\n\n供局域网发现使用，不包含配置、地址或凭据。",
        "security": [],
        "x-required-scope": "public"
      }
    },
    "/api/v1/hardware": {
      "get": {
        "summary": "获取硬件连接状态",
        "operationId": "hardware_status",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HardwareInventory"
                }
              }
            }
          },
          "400": {
            "description": "Host 不被允许。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "Host 不被允许。"
                }
              }
            }
          },
          "401": {
            "description": "凭据缺失、无效或已过期。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "凭据缺失、无效或已过期。"
                }
              }
            }
          },
          "403": {
            "description": "权限不足，或浏览器来源 / 校验头不符合要求。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "权限不足，或浏览器来源 / 校验头不符合要求。"
                }
              }
            }
          }
        },
        "tags": [
          "Hardware"
        ],
        "description": "**最低权限：read。**\n\n返回候选设备、各部件状态及连接配置。扩展板不提供风扇和灯条的插接反馈，应结合实际安装情况配置。",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-required-scope": "read"
      },
      "put": {
        "summary": "更新硬件连接配置",
        "operationId": "hardware_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HardwareCommand"
              },
              "example": {
                "expected_revision": 1,
                "value": {
                  "cube": "auto",
                  "oled": "auto",
                  "case_fan": true,
                  "lights": true,
                  "screen": true
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HardwareInventory"
                }
              }
            }
          },
          "422": {
            "description": "参数、请求体结构或业务字段校验失败。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "参数、请求体结构或业务字段校验失败。"
                }
              }
            }
          },
          "400": {
            "description": "Host 不被允许。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "Host 不被允许。"
                }
              }
            }
          },
          "401": {
            "description": "凭据缺失、无效或已过期。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "凭据缺失、无效或已过期。"
                }
              }
            }
          },
          "403": {
            "description": "权限不足、跨来源写入，或 Cookie 请求缺少 X-Pi-Control: 1。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "权限不足、跨来源写入，或 Cookie 请求缺少 X-Pi-Control: 1。"
                }
              }
            }
          },
          "413": {
            "description": "声明的请求体超过 32 KiB。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "声明的请求体超过 32 KiB。"
                }
              }
            }
          },
          "409": {
            "description": "配置版本冲突、部件不可用、操作过于频繁或资源限制；请阅读 detail。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "配置版本冲突、部件不可用、操作过于频繁或资源限制；请阅读 detail。"
                }
              }
            }
          }
        },
        "tags": [
          "Hardware"
        ],
        "description": "**最低权限：control。**\n\n使用本接口的 `expected_revision`，不是控制配置的 `expected_version`。先 GET `/hardware` 读取 revision，再提交完整 value。`auto` 自动连接，`off` 不连接，也可指定扫描结果中的设备 id。停用后停止管理并保留物理状态。",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-required-scope": "control",
        "parameters": [
          {
            "name": "X-Pi-Control",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "1"
              ]
            },
            "description": "请求带有网页登录 Cookie 时，写操作必须提供；仅 Bearer 认证时无需此头。"
          }
        ]
      }
    },
    "/api/v1/hardware/scan": {
      "post": {
        "summary": "重新扫描兼容部件",
        "operationId": "hardware_scan",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HardwareInventory"
                }
              }
            }
          },
          "400": {
            "description": "Host 不被允许。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "Host 不被允许。"
                }
              }
            }
          },
          "401": {
            "description": "凭据缺失、无效或已过期。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "凭据缺失、无效或已过期。"
                }
              }
            }
          },
          "403": {
            "description": "权限不足、跨来源写入，或 Cookie 请求缺少 X-Pi-Control: 1。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "权限不足、跨来源写入，或 Cookie 请求缺少 X-Pi-Control: 1。"
                }
              }
            }
          },
          "413": {
            "description": "声明的请求体超过 32 KiB。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "声明的请求体超过 32 KiB。"
                }
              }
            }
          }
        },
        "tags": [
          "Hardware"
        ],
        "description": "**最低权限：control。**\n\n检查配置允许的 I²C 总线及兼容地址，并尝试恢复部件。请求完成不等于所有部件可用，请读取 components 和 scan_errors。",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-required-scope": "control",
        "parameters": [
          {
            "name": "X-Pi-Control",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "1"
              ]
            },
            "description": "请求带有网页登录 Cookie 时，写操作必须提供；仅 Bearer 认证时无需此头。"
          }
        ]
      }
    },
    "/api/v1/devices": {
      "get": {
        "summary": "列出附近及已记住的设备",
        "operationId": "devices",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDirectory"
                }
              }
            }
          },
          "400": {
            "description": "Host 不被允许。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "Host 不被允许。"
                }
              }
            }
          },
          "401": {
            "description": "凭据缺失、无效或已过期。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "凭据缺失、无效或已过期。"
                }
              }
            }
          },
          "403": {
            "description": "权限不足，或浏览器来源 / 校验头不符合要求。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "权限不足，或浏览器来源 / 校验头不符合要求。"
                }
              }
            }
          }
        },
        "tags": [
          "Devices"
        ],
        "description": "**最低权限：read。**\n\n返回本机发现缓存和服务模式；记住的离线设备会保留，并定期检查。Docker 演示服务也能记住目标地址，选择后进入目标控制台。",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-required-scope": "read"
      }
    },
    "/api/v1/devices/scan": {
      "post": {
        "summary": "刷新设备发现",
        "operationId": "scan_devices",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDirectory"
                }
              }
            }
          },
          "400": {
            "description": "Host 不被允许。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "Host 不被允许。"
                }
              }
            }
          },
          "401": {
            "description": "凭据缺失、无效或已过期。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "凭据缺失、无效或已过期。"
                }
              }
            }
          },
          "403": {
            "description": "权限不足、跨来源写入，或 Cookie 请求缺少 X-Pi-Control: 1。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "权限不足、跨来源写入，或 Cookie 请求缺少 X-Pi-Control: 1。"
                }
              }
            }
          },
          "413": {
            "description": "声明的请求体超过 32 KiB。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "声明的请求体超过 32 KiB。"
                }
              }
            }
          },
          "409": {
            "description": "配置版本冲突、部件不可用、操作过于频繁或资源限制；请阅读 detail。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "配置版本冲突、部件不可用、操作过于频繁或资源限制；请阅读 detail。"
                }
              }
            }
          }
        },
        "tags": [
          "Devices"
        ],
        "description": "**最低权限：admin。**\n\n通过本机 mDNS 发现数据检查其他 Pi Control 实例，并探测已记住地址。扫描正在进行时返回 409。",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-required-scope": "admin",
        "parameters": [
          {
            "name": "X-Pi-Control",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "1"
              ]
            },
            "description": "请求带有网页登录 Cookie 时，写操作必须提供；仅 Bearer 认证时无需此头。"
          }
        ]
      }
    },
...
    "/api/v1/oled/frame": {
      "get": {
        "summary": "获取最近写入的屏幕帧",
        "operationId": "frame",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/OledFrame"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Host 不被允许。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "Host 不被允许。"
                }
              }
            }
          },
          "401": {
            "description": "凭据缺失、无效或已过期。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "凭据缺失、无效或已过期。"
                }
              }
            }
          },
          "403": {
            "description": "权限不足，或浏览器来源 / 校验头不符合要求。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "权限不足，或浏览器来源 / 校验头不符合要求。"
                }
              }
            }
          }
        },
        "tags": [
          "OLED"
        ],
        "description": "**最低权限：read。**\n\n没有成功写入记录时返回 JSON `null`。画面是最近成功发送的位图，不是摄像头采集。",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-required-scope": "read"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceIdentity": {
        "additionalProperties": true,
        "properties": {
          "id": {
            "description": "设备持久化 UUID。",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "version": {
            "description": "服务版本。",
            "title": "Version",
            "type": "string"
          },
          "product": {
            "const": "pi-control",
            "title": "Product",
            "type": "string"
          },
          "protocol": {
            "const": 1,
            "title": "Protocol",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "version",
          "product",
          "protocol"
        ],
        "title": "DeviceIdentity",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "在设置中创建的 pc_ 令牌。Authorize 输入完整令牌，无需 Bearer 前缀。"
      },
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "pi_control_session",
        "description": "通过 /auth/login 创建的网页登录会话，浏览器自动携带。写请求同时需要 X-Pi-Control: 1。"
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Your Pi Control service"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "SessionCookie": []
    }
  ],
...
}